backend/imbedding/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
redis==5.0.1
python-dotenv==1.0.1